Add files via upload
So here you can change the color of the columns via tapping 'space'. You can also "lock a color" if you tap on the lock icon. Every color name (#000000) you can copy. You can also share a collection of chosen colors because the names of colors save in the hash. 

http://127.0.0.1:5500/index.html